+++ RELEASE_NOTES.md
@@ -5,6 +5,9 @@ Release Notes
 * Container Apps: Support for collections of env vars, fix ACR credentials linking.
 * Container Apps: Add Dapr / App Insights integration.
 
+## vNext
+* Virtual Machines: Accelerated networking and IP forwarding for interfaces
+
 ## 1.7.15
 * Adds the West US 3 location
 * Virtual Machines: Support for multiple network interfaces on a VM